Revert "move stuff to utils"
This reverts commit 92d125a7d7049dd4684f54e3196695bf083d292b.
